split increase deployment cost test
