+++ integration-tests/src/tests/client/features.rs
@@ -6,6 +6,7 @@ mod cap_max_gas_price;
 #[cfg(feature = "protocol_feature_fix_contract_loading_cost")]
 mod fix_contract_loading_cost;
 mod fix_storage_usage;
+mod increase_deployment_cost;
 mod lower_storage_key_limit;
 mod restore_receipts_after_fix_apply_chunks;
 mod wasmer2;
